moved Details and Reviews API to seperate compnients
