+++ src/App.css
@@ -122,8 +122,8 @@ header {
 	display: block;
 	object-fit: cover;
 }
-/* ViewRestaurant */
-.View-Restaurant-images{
+/* ViewRestaurant RestaurantsDetail */
+.RestaurantDetail-images{
 	display: flex;
 	flex-direction: row;
 	align-items: center;
@@ -131,7 +131,15 @@ header {
 	
 }
 
-.View-Restaurant-images img{
+.RestaurantDetail-images img{
 	border: 2px solid darkblue;
 	border-radius:10px;
+}
+
+
+.Restaurant-schedule{
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	justify-content:space-between;
 }
