+++ src/Components/PickRestaurants/ViewRestaurant/AddFavorites/AddFavorites.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const AddFavorites = () => {
+  return (
+    <div>AddFavorites</div>
+  )
+}
+
+export default AddFavorites
